# Avoid string allocations in BacktraceCleaner